backend/routes/wallet.py
from fastapi import APIRouter, HTTPException
from xrpl.wallet import Wallet
from xrpl.clients import JsonRpcClient
from pydantic import BaseModel

router = APIRouter()

class WalletRequest(BaseModel):
    seed: str = None

@router.post("/create")
async def create_wallet(request: WalletRequest = None):
    try:
        wallet = Wallet.create() if not request or not request.seed else Wallet(seed=request.seed)
        return {
            "address": wallet.classic_address,
            "seed": wallet.seed
        }
    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))

@router.get("/balance/{address}")
async def get_wallet_balance(address: str):
    try:
        client = JsonRpcClient("https://s.altnet.rippletest.net:51234/")
        # Placeholder for balance retrieval logic
        return {"address": address, "balance": "0 XRP"}
    except Exception as e:
        raise HTTPException(status_code=400, detail=str(e))
