requirements.txt
xrpl-py==2.3.0
web3==6.0.0
pillow==10.2.0
python-dotenv==1.0.0
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
alembic==1.10.3
psycopg2-binary==2.9.6
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
requests==2.28.2
gunicorn==20.1.0
ipfs-http-client==0.7.0
cryptography==42.0.2
